Utilize bits for short strings instead of integers
The basic rationale is our use of POS/NEG doesnt give us a big
enough range in the positive integers (effectively -9..32) to make
it worth while to allocate so many bits to them. So we make the
bits available to make it possible to represent longer strings
with a single byte/type length marker.

Further changes to how we represent "common" numbers will follow.