Hash dumping
Hash keys are strings and thus still todo, but the abstract "dump a
hash" function is complete unless we were to decide to serialize all
keys followed by all values. But that's expensive in the encoder (Perl
hash API does not work that way) or needs to be an option.

Hey, this implementation stuff is fun and quick progress!