MEMDEBUG option to Makefile.PL: disable optimistic preallocation
By default, we grow the buffer size geometrically and pre-allocate
rather more than less. By invoking Makefile.PL as follows, the build
will disable the initial buffer preallocation of 16k as well as the
geometric growth. This is for debugging invalid writes.

  MEMDEBUG=1 perl Makefile.PL

In detail, Makefile.PL will propagate this to -DMEMDEBUG during the
build stage. The same works for DEBUG, by the way, which enables debug
symbols during the build.