Make shared hash keys optional (but default)
Also, this commit introduces what I thought to be a significant
optimization. It's disabled by default because it violates perl's
encapsulation big time. Also, it doesn't really help as much as I
expected. This could do with some debugging. Otherwise, we might take it
back out altogether.

The optimization is to check the refcount of shared hash keys. If
not > 1, then we should not even bother tracking them in the string ptr
table. Not measureable in our rather shared-string-heavy benchmark
script.

This and other, later hacks that might be unsafe / non-portable /
non-backwards-compatible can be enabled in via ENABLE_DANGEROUS_HACKS=1
as an ENV variable at Makefile.PL time.