rework the logic for handling weakrefs
This reworks how we handle weakrefs. Now the weak_seenhash stores a map
of referent => weakref-ofs, or 0 if a normal ref has been seen for that
referent.

When we encounter a weak referrent:

    If we have not seen it before we store 0
        if we are dumping a weakref we store the ofs of the weakref tag.
        otherwise we are dumping a normal ref so we store 0
    otherwise if we have seen it before and we are dumping a normal ref
        we update the value and set it to 0.

Then at the end every non-zero weakref ofs stored in the hash has its
corresponding weakref converted to a PAD op, turning it into a normal ref.

This ensures that when we deserialize every ref in the structure has a
refcount of 1 or more.