eliminate cross contamination during benchmark
If a serializer calls something like SvPV(foo) the object will be
upgraded to a PV type, which may impact how other serializers would
serialize it.

For instance message-pack and sereal, and to an extent all of the
serializers are subtly sensitive to the real type of a scalar. Storable
on the other hand seems to call SvPV(foo) on everything. This means
that if mp or sereal see the strucutre after Storable has looked at it
they will see something that is quite different than had they seen it
before.

This patch changes the benchmark so each serializer operates on their
own private copy of the structure. The results on serialization sizes
and speed are drammatic.