fix iteration logic
I noticed that once in a while the PAD logic would fail. Reviewing
the iteration logic it seems to me that the existing code would never
look at items in the last bucket, as right after we fetch the entity
for the last bucket we would reset it back to null.

IOW the logic was:

    loop
        fetch the top entity from the current bucket
        increment the bucketnumber
        if the new bucket value is larger than the allowed buckets
            reset the top entity to null
            exit loop
    while entity is null

Which makes it more clear that we wont use the entity chain in the
last bucket.

This changes the logic to

    loop
        if the current bucket value is larger than the allowed buckets
            reset the top entity to null
            exit loop
        fetch the top entity from the current bucket
        increment bucket number
    while entity is null

Which will use the entity if it exists by exiting the loop before it
checks the next bucket number.

It would be nice if we had some tests for this logic.