split up t/010_desperate.t so the test set can be shared between the encoder and decoder
This moves the test sets into a shared file

  t/test_set.pl

which is set up as a symlink to

  shared/t/test_test.pl

by the Makefile.pl

This should make it easier to keep the two repos in synch as we continue to dev.

This also adds two ignores, 'dt' and 'et', into the Encoder and Decoder
respectively, so you can add a symlink to the other repos tests under
these names if you wish. I find this useful in cross checking