OO interface: 40% faster for tiny data structure
The OO interface can be 40% faster for tiny data structures since it
reuses the encoder struct and doesn't need hash munging for options
every time. It also doesn't free the encoder buffer, so that may help if
we're encoding many similar data structures.

For the large benchmark case, it seems to improve consistently by
approx. 2%. Not much, but not nothing.

Conflicts:

	Sereal-Encoder/author_tools/bench.pl
	Sereal-Encoder/srl_encoder.c
	Sereal-Encoder/t/010_desperate.t