Speed up functional interface by buffer stealing
The encoder buffer can be safely stolen in the functional interface. In
the OO version, we may be reusing the buffer, so a malloc cannot be
avoided, but there, the better trade-off seems to malloc only what's
necessary for the final output instead of nuking the encoder buffer AND
using more memory for the output than necessary.

This makes the functional interface (in the absence of encoder options)
virtually as fast as the OO interface - in the current large-structure
benchmark.