More documentation for Encoder and Decoder
Encoder isn't too bad now (including documentation for all three public
functions), but the Decoder docs need a lot of work.

We'll have to think carefully about how to implement protocol-level
compatibility and how to expose it to Perl. Also, should we have a
protocol spec exposed as POD somewhere?