Avoid resetting persistent decoder flags
Some (most, right now) flags in the decoder are "static", IOW, they are
not intended to be reset for each decoder run. For example, one of the
flags marks the difference between a decoder that will be freed soon
versus a decoder that will be reused.

We were unconditionally unsetting all flags in various cleanup places.
That was bad. Instead, we just unset all flags that pertain to the
current decoder run.