Always call av_extend - the 8 default size applies to hashes only
We've been assuming that by default, arrays have an underlying C array
of 8 SVs, but that's not true, as far as I can tell. sv.c has that logic
for hashes, but not for arrays. So we always call av_extend - even
though it could be argued that it shouldn't be called for length=0
arrays, but the extra branch is probably not worth that comparatively
rare special case.