Improve header, reduce number of "tags" allocated to ASCII strings.
This defines a new header and changes some of our tag allocations.

As of this release the magic string for serial is four bytes consisting
of "=srl". The '=' byte is reserved as a tag so that it will never be
used as a tag inside fo the dump.

Additionally we now only use 32 tags for short strings, so we can
reallocate the other 32 to hash and array references.