Implement (stringify|undef|warn)_unknown options to the Encoder
Basically, if we find unknown data structures (references, really), we
have several options: die (default), emit undef, or stringify the data
structure. If either of the latter two, we optionally emit a warning,
too.

Not being too strict with your input data makes a lot of sense if you're
working on the system that is used to track problems. You never want
that flat-out dying rather than just making do.