change SRL_INLINE to SRL_STATIC_INLINE and define it properly on Win32
These terms are not portable. Win32 for instance considers "inline" to
be a post C89 addition and as such it is spelled __inline. Perl defines
various macros for this purpose, so we could use them instead and avoid
any portability issues - except that PERL_STATIC_INLINE *isn't* defined
as far as I can tell. Which means we have to fake our own. This might
need more refinement for exotic platforms, and etc.

Many thanks to bulk88 for the Win32 guidance.