Number offsets from 1 in relocatable format, not 0
This preserves the truthiness properties of the offset while at the same time
making it relocatable, and avoids a lot of awkward logical changes that
would result from doing it otherwise.

Getting the tests to the right thing was kindof a bodge. I didnt have
time to do it properly.