Remove spurious SvREFCOUNT_dec
This appears to be the correct resolution of the

  Attempt to free unreferenced scalar: SV ...

problem. We use hv_common with the LVALUE flag to fetch an entry
and create it if it doesn't exist. At least if it was freshly created,
the SV will have a refcount of 1. Not sure what the behaviour is
otherwise, but I believe that hv_common doesn't increment refcounts for
you for existing entries. This fetch was followed by a refcount
decrement (in some circumstances only!) and that freed the SV. But we
continued to use the SV. Bang.