Perl: Make Encoder depend on Decoder (2.00 and up)
This is kind of a big change: Previously, one could install encoder and
decoder independently freely. This makes the encoder depend on a
decoder. The rationale is that in a major upgrade cycle, one always
needs to upgrade the decoder FIRST anyway. And an encoder without a
decoder is rarely useful. What we get by requiring the decoder here is
that more users will be able to run more tests, so that both they and we
(CPAN testers) get more insight into where Sereal works and doesn't.

So everybody wins. I hope.