Spec: Proposed changes to implement freeze/thaw hook mechanism
Make sure to compare against the logic for CBOR hooks which is
sufficiently generic to serve multiple serializers. Right now, the only
incompatibility is that for Sereal, FREEZE needs to return a single data
structure instead of a list. That is quite a bit more efficient for
simple data structures and also easier on the implementation,
particularly for more static languages than Perl.

Comments welcome. This is a proposal.