Perl: Update freeze/thaw benchmark
This adds a serialization case of objects without FREEZE hook but with
the FREEZE/THAW feature enabled. This is the cb_nocbdata case below:

                              Rate     cb cb_nocbdata  no_cb no_cb_nocbdata
  cb             17886.5+-0.0016/s     --      -72.8% -90.2%         -90.3%
  cb_nocbdata       65716.4+-4.1/s 267.4%          -- -63.9%         -64.3%
  no_cb                181839+-0/s 916.6%      176.7%     --          -1.1%
  no_cb_nocbdata   183939+-0.055/s 928.4%      179.9%   1.2%             --

This means that with the freeze_callbacks option set, simply attempting
to look up the FREEZE method in a class repeatedly can be a significant
slowdown. It remains to be seen whether a custom hash based cache for
"class name => hook (or NULL)" would be faster than the method lookup.
That cache isn't implemented yet, would potentially introduce bugs in
weird corner cases (callbacks modifying FREEZE methods, WTF?) and is
generally somewhat unlikely to speed things up.

So this is likely it: If you use freeze callbacks, you pay for it a
little bit. (This is a worst-case benchmark.) But that's ok. Because you
don't have to and because all other serializers that offer this
interface are A LOT slower still. :)