Perl: Tentative fix for FREEZE and referential integrity in ENcoder
The decoder is still broken. And this is not clearly
a complete fix. It might be that I am inadvertently breaking
ALIAS semantics for scalar refs to objects. *sigh* What a twisted
mess of semantics.

Furthermore, I am not very happy with the way this code works
wrt. adding a new hash instead of using the normal ref-seen-hash.
Alas, using that seems hard due to the special FREEZE logic.