go: cache struct tags and create if no tags present
When repeatedly processing the same type, we now cache the parsing of
the struct tags.  This is a considerable speed-up.  We also construct a
name->index mapping to eliminate the special casing of having
struct-tags or not.  (If there are no tags, we construct a mapping from
the type's exported fields.)