ruby: add REFP support in the encoder
Sereal.encode(object,Sereal::REF)
or Sereal::REF|Sereal::SNAPPY_INC, or Sereal::REF|Sereal::SNAPPY

when encoding will try to use Sereal's REFP tag to transmit only the
the original object's offset in the packet.
So:
   one = [ 1,2,3,4,5 ]
   two = [ one, one ]
   Sereal.encode(two,Sereal::REF)
will send 'one' only once, and one REFP that points to the first one
it uses one.object_id as a hash key in a local tracker hash
and if it sees this object_id again it just sends the offset.