Make sure that OBJECT_FREEZE "payload" is *always* REFN ARRAY and never ARRAYREF
If we were to allow FREEZE to serialize items as NON objects then we want the goto.

And it would fix the problem this fixes. However it would introduce other issues for this FREEZE api.

So for now we do some nasty stuff to trick srl_dump_av() into dumping this as REFN/ARRAY (as it would
with >1 refcount anyway) but artificially incrementing the "refcount" on the array