Fix issue with OBJECT items that use COPY for the class name
Damians GO implementation creates OBJECT items with the class name
defined via a COPY statement. The Perl implementation does not do this, so
we never noticed the stupid bug where we were looking up the stash data from
the ref seen hash (ref_seenhash) instead of the similarly named stash seen hash
(ref_stashes). Because of the similarity in the name it took me ages to spot
the problem. :-(

Probably a follow-up patch should rename one of the two hashes to something
very distinctive so such bugs in the future are less likely.

This needs tests. We have no tests for this functionality currently.