Perl: Fix reentrancy bug in the encoder
This pertains to RT ticket #93560 as reported by Zefram.

If in a FREEZE hook, the same encoder object is reused, up to this
commit, we'd emit corrupted Sereal for the INNER encode call.

There is logic to detect reuse of a dirty encoder and to lazily
clone the dirty encoder without the encode state. That logic is actually
mostly working (see minor changes in the srl_build_encoder_struct_alike
function). The bug was that this newly allocated, virgin encoder
structure would be used for the inner encode run alright, but the XS
wrapper code that extracts the output string was oblivious to the fact
that its encoder struct was not the one that was used for the encode.
The solution is to always return a pointer to the encoder struct that
was finally used for the encode. Maybe there's more elegant solutions,
but this one is simple and efficient.