Speed up dumping hashes
The backreferences info is stored in the HvAUX struct, which is lazily
created by perl in an expensive operation that requires realloc()ing
the bucket array. When this struct is created the hash is marked
SvOOK().

Since we weren't checking for this flag we were effectively realloc()ing
nearly every hashes bucket array during dumping, with the associated
speed costs.

Thanks to Steffen for noticing the slowdown and asking me to look into
it.