s/sereal_(decode|encode)(\w*)/sereal_encode${1}_op/g
Rename the functional/custom opcode interface to end with an _op suffix.
The whole encode_sereal/sereal_encode dichotomy is to likely to lead to
confusing. Adding a suffix makes it clearer what is going on.

My choice of suffix was more or less "_op" or "_func" I went with "_op".