Framework for a potential libsereal
The intention of this is to have a pure C implementation of Sereal for
people to learn (and copy) from. The Perl/XS reference implementation is
very, very cozy with the Perl internals and so is hardly accessibly to
those who haven't studied those to an extent.

The additional challenge is that C obviously comes with no dynamic data
structures and/or inspection. That means this is also a useful
experiment in APIs.

This is considered experimental and is likely never going to be a full
reference implementation.

Already comes with ptr- and string hash table implementations.