Perl: Rework how we dump simple SV's for consistency and better representation.
This prefers integers/floats over strings, except in a special case
where IOK/NOK/POK, which seems to only occur when handling engineering
notation like "0e0" and similar constructs. As these types of strings
can carry extra meta data (such as "0e0" == "0e1" but "0e0" ne "0e1")
outside of their IV representation we still send them as strings.

As part of this the logic for handling simple SV's was refactored (via
ugly macros) so that it is consistent across several callsites.
Previously it was not consistent.