Deal with SvPVLV from tied arrays properly by respecting private flags
It turns out that tied arrays in older perls return
SvPVLV's with various private flags like IOKp, POKp, NOKp set
which when followed "just work".  So when we changed the logic
to start respecting IOK, POK, NOK (the "public" flags), we started
failing to handle SvPVLV's properly, breaking the encoder on
various older perls.

So we add that logic back in *after* we honor the public flags.

Sigh.