Go: Rework compat test skipping
It's known and expected that some of compat tests fail. It's due to a
deliberate way the Go implementation is done (dgryski can elaborate on that
more).

But the way these supposed-to-fail tests were checked was way suboptimal:
literally just checking a number of failed tests.

This commit explicitly lists names of all tests that are known and ok to fail,
and we simply skip them, producing proper TAP output.