go: don't treat COPY tag as an offset
In isStringish, we were forgetting to increment idx, which means that
varintdecode pulled the COPY tag out as the source offset for the COPY.
In our tests, we had string data at offset 0x2f, which meant it
satisfied the stringish check.  The *actual* string parsing code was
correct, so we would extract the right class name.

The hash randomization change in 1.3 changed the order that struct
fields were processed (since we stuff the fields in a map), causing
offset 0x2f to sometimes no longer point into the middle of a lower-case
string that looked like a SHORT_BINARY tag.