Go: Allow to specify a protocol version for a test corpus
On command line you can pass to make(1) a Sereal procotol version to use for
generating a test corpus, for instance:

    $ make compat CORPUS_PROTO_VER=3

Default version remains 2, for the time being.

The change has one negative implication though: test_dir is re-generated every
time the compat tests are run. This is because my make-fu is weak and I want to
ensure test_dir always correspond to a version specified via CORPUS_PROTO_VER.