Go: Add support of the new CANONICAL_UNDEF tag added in v3
Sereal protocol version 3 introduced the new CANONICAL_UNDEF tag, used to
represent Perl's canonical shared undefined value (PL_sv_undef) in certain edge
cases. For more details see:
https://github.com/Sereal/Sereal/blob/master/sereal_spec.pod#user-content-dealing-with-undefined-values

The implementation re-uses the existing type PerlUndef. In Go there is no
distinction like there is in Perl, hence it makes sense to have all flavors of
undef to be of the same type. However for compatibility with Perl we provide a
way to pick which exact undef you want to encode.

Consider --ignore-all-space to view the diff.