Go: Make compressor work with opaque data; refactor MarshalWithHeader()
This is a follow up on 16cad24. Compressor is supposed to know nothing
about a document header and should work with a raw payload only. This
commit is the first step toward the goal.

Now a compressor takes an opaque blob of data, compresses, and
returns it back.

As a side change I renamed variables b and encoded; and explicitly
separated encoded document header, optional header suffix, and body.

Concatination of encoded header and body is done in a suboptimal way,
but I plan to improve it later on.