rework our deep comparison tests
I want to reduce the number of tests we produce, but get better diagnostics
at the same time. This can be done by turning the bulk tests into a single
"test", which is facilitated by using "fail" if we have an issue, and
issuing a pass() when we have finished all our tests successfully.

At the same time we switch to using our own deep comparison logic, so we
can track aliases, check utf8 flags, and do other types of comparison
not handled by Test::Deep/Test::More.