Sereal::Merger: implement parser stack as simple array
Originally, I wanted to use perlapi for dynamically growing arrays.
But av_push(), av_pop() and other functions popped up in my profiling
which meant that either they are not as fast as it could be
or I don't know how to cook them. I decided to go first way :-).