Sereal::Merger: make caler fill fields in STRTABLE_entry
Profilling shows that BODY_POS_OFS() macro is quite expensive when
it called millions of times. So, by letting the caller be responsible
of filling in tag_offset field I safe on BODY_POS_OFS() call when there
is already duplicate in strtable.

One more thing is that this step is an easy one because the caller is
already responsibe for setting str_offset field.