proto5: By default do not use LONG_DOUBLE, and fix various related issues
Unlike float and double precision, long double is platform and build
option dependent. For instance on a 32 bit Intel box long double is
stored as a 12 byte 80 bit value, and on 64 bit Internal it is stored
as a 16 byte 80 bit value. There could be other definitions.

We therefore deprecate use of SRL_HDR_LONG_DOUBLE as it is non portable
and at the same time define it to always be 16 bytes long.
At the same time we disable the encoder from producing it without
the special build option SRL_ALLOW_LONG_DOUBLE.

At the same time this means on USE_LONG_DOUBLE perls we have percision
loss issues in our tests which necessitated some adjustments.