newSV() is faster than newSV_type(SVt_NULL)
I accidentally use newSV(SVt_NULL), which is the same
as newSV(0), instead of using newSV_type(SVt_NULL), and saw
a speedup around 5% in our benchmarks. Inspecting the code,
newSV() does not call sv_upgrade() when the length is 0,
whereas newSV_type() *always* calls sv_upgrade, even if the
type is SVt_NULL.

This can account for as much 5% speedup in the bench test.