proto4: first steps towards a functional SRL_HDR_MANY tag
This implements the IVDRLE MANY subtype. This is a run-length
encoded delta representation of a sequence of integers. It packs
repeated, and incrementing sequences of integers very densely,
and produces compact output for sequences of large integers
centered around a common mean.

Not only does this produce smaller dumps, but when it does,
it does it MUCH faster too. Both on decode and encode.

This suggests there are a lot of improvements available if we
can speed up processing of arrays of simple types.