Perl decoder: In buffer length sanity check, be more robust
The sanity check for buffer length remaining used an unsigned value for
the pointer difference of end-pos. That should never be negative, of
course, but if a prior bug caused corruption in the decoder/buffer state
already, we'd continue along and segfault happily. Instead, we're now
using a ptrdiff_t which should convert this happy segfault in an even
happier Perl exception.

This won't fix the underlying bug that got you to this state, so it's
not a complete bug fix for what we're hunting. But it's good defensive
programming.