Fix decoding of FROZEN objects
We expect exactly a REF to an ARRAY after a FREEZE.
There's probably other representations that could come through from other languages,
but I need use-cases for them, and I'll follow YAGNI for now.

This fix probably only fixes one use-case, but the tests pass,
and that's a start. If anyone wants to add other tests that fail,
I'm happy to look at fixing them.