Portability fix for Mac OS X
It appears that on Yosemite, __GNUC__ is defined, but not __GLIBC__.
Therefore 11a20ffa broke the build by making the __GNUC__ block
attempted to compile in csnappy_internal_userspace.h ; while the
right one was the __APPLE__ block, obviously. The solution is
to reorder the #define tests from the most specific to the least
specific.