Fix nil pointer reference encoding
Two references to nil pointers of different types would cause
the second one to be a REFN to the first.
On decode that causes us to try putting a struct of a different type in the same place,
and we get an error.

This fix is pending tests in Sereal, since it was fixed against our own tests.