Deal with leading zeros and non-canonical forms when [IN]OK + POK
If a string like "0000" is used in numeric context perl will treat it as
a raw 0, but we do not want to serialize it as a 0, just in case the leading
digits are significant (which they are in various contexts, like boolean).
Similarly if there is a string like "         1" and it is used in numeric
context, it will be IOK/POK, with an IV of 1, but obviously not be the same
as "1".

A similar issue comes up with a string like "0.0", which is true in string
context, and false in numeric context. If we serialize it as the NV 0,
then when perl stringifies it later it will come back as "0" which is false
in string context. Thus changing the values meaning.

We use a simple heuristic to decide what to do. If the value is POK and
IOK/NOK, then we check the numeric value of the string, if it is
numerically non-zero then we check that the first char of the string is
'1' to '9', if it is not then we emit the string form, otherwise as the
numeric type. (This handles sane cases, like "    1" and ".1" and "00001",
but is not  reliable for arbitrary dual-vars like produced by
Scalar::Util::dual_var). On the other hand if the value is numerically false
(equals zero), then we check the length of the string, if it is longer
than 1 char long then we output the string value, if it 1 char long then
we emit the numeric value (0).

Note that for a string like "10foo" perl does not set the IOK flag. It
only sets the pIOK flag, so it does not cause problems here.