add TODO tests for various "esoteric" numeric/string cases
Strings like "0.0" lose their value when they have been used in
numeric context, because they get serialized as NV's when their
string value is not "canonical" for their NV value (0). Similar
issues with strings like "    1    " which will have a numeric
value of 1, but the string value cannot be recreated from the
numeric value. This adds a handful of cases which might be affected.

A future patch will fix this.