rework preallocation logic a bit
Make it easier to debug/visualize preallocation patterns
for arrays, hashes, and overall.

In particular instead of doubling the buffer, we use a slightly
different model, using (cur_size + (req_size >> 2)) as a new
ceiling. Which is close to using N *= 1.25.

We also standardize the rules for estimating size requirements
for hashes and arrays by switching from a conservative model
to a more liberal model.