Perl: Add accessor macro that doesn't reify
The SRL_GET_WEAK_SEENHASH macro will automatically reify the ptrtable
for weak references on use. This is convenient and since many data
structures contain no weak-refs, this means that since
SRL_GET_WEAK_SEENHASH is called only within the weak-ref branch, we
avoid needlessly creating a ptrtable in the encoder object.

Wait.

SRL_GET_WEAK_SEENHASH is called unconditionally in srl_fixup_weakrefs.
And it's called to check whether we encountered any weak-refs. That
means we're avoiding the instantiation of the ptrtable all the way (and
in the cases of needing it, we're branching for each use), but in the
end, we instantiate anyway if we didn't need it thus far.

That's kind of dumb. (I wrote that code.)

So this change simply avoids reifying that thing at the last moment when
it's not necessary. Shouldn't be a significant optimization, but still.