Fix Issue #82 - blessedness lost from unreferenced scalars
The rewrite mechanism which handles REFP was breaking
blessed scalar refs which were referenced as scalars prior
to being referenced as a ref. The way perl manages objects is
that the referent is blessed, but only a *reference* to a blessed
scalar is usable as an object. Which means that unreferenced scalars
should NOT be output as objects, instead the first reference to
a scalar should be blessed. (This applies to hashes and arrays as
well, but you cant have an unreferenced array/hash in a data structure
constructed from perl, so it doesn't come up.)

The simplest solution to this is to track which items we have output
as objects, and do some better book keeping for our ref rewrite logic.
One side effect of this is we really need to know what type of SV's
we cannot handle, so that we dont try to bless non refs in the logic
that handles unsupported types.