rework sort order so it can be more efficient
This changes the canonical order - new order is shortest to longest, then
a byte comparison on the raw representation of the string, then by utf8ness
with non-utf8 preceeding utf8.

The idea of this sort order is it can be made fast, and it can be used
on all types of hashes, with special cases only for the 128 code points
that are "was-utf8" flagged in hashes. (Was-utf8 refers to the fact
that Perl should upgrade the key to utf8 even though it is stored as
latin-1/binary in the HE structure).