use hardlinks on Win32 instead of symlinks to make repo buildable
Previously the git repo couldn't be built on Win32 because symlink throws
an exception that is it unimplemented on Win32 perl. Hard links are
implemented on Win32, and usually available since nearly all modern Win32
OS installs use NTFS (if you keep Sereal git repo on a USB stick this will
fail). Use a constant for constant folding since the sub is called in a
loop.