Non-glacial speed
With this logging granularity, the cost of computing the logging string and then
discard it is way too high, and it also affects the case when logging is
disabled.

Since this is meant for developer debugging anyway, replace the log4j-y interface with
an ad-hoc tracing boolean. I'm not even sure this level of tracing is needed,
but at least now it takes seconds to run TestCopus rather than minutes.